refactor: Rename `ResetPassword` > `ResetPasswordPage`
